* in 1.3: Changed types used in rrd_file_t from off_t to size_t to prevent: warning: comparison between signed and unsigned * Misc other off_t -> size_t changes to prevent signed warning * in 1.3: Added rrd_dump_opt_r declaration to prevent undef warning * Explicit cast to time_t of check in rrd_fetch to prevent signed warning * Added vasprintf path to sprintf_alloc to prevent excessive mallocing * in 1.3: fixed STRERROR_R_CHAR_P undef warning
patch by Steven Hartland killing multiplay.co.uk




git-svn-id: svn://svn.oetiker.ch/rrdtool/branches/1.3/program@1808 a5681a0c-68f1-0310-ab6d-d61299d08faa

+++ src/rrd_open.c
@@ -310,7 +310,7 @@ rrd_file_t *rrd_open(
       for (i=0; i<rrd->stat_head->rra_cnt; i++)
         row_cnt += rrd->rra_def[i].row_cnt;
 
-      off_t correct_len = rrd_file->header_len +
+      size_t correct_len = rrd_file->header_len +
         sizeof(rrd_value_t) * row_cnt * rrd->stat_head->ds_cnt;
 
       if (correct_len > rrd_file->file_len)
@@ -389,10 +389,10 @@ void rrd_dontneed(
     rrd_t *rrd)
 {
 #if defined USE_MADVISE || defined HAVE_POSIX_FADVISE
-    unsigned long dontneed_start;
-    unsigned long rra_start;
-    unsigned long active_block;
-    unsigned long i;
+    size_t dontneed_start;
+    size_t rra_start;
+    size_t active_block;
+    size_t i;
     ssize_t   _page_size = sysconf(_SC_PAGESIZE);
 
     if (rrd_file == NULL) {
